+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Eliminated console warnings on application load
 - Schema panel on Data tab always show dataset schema
 - Minimized usage of `!important` directives in CSS
+- Fixed bug with empty offset interval
 
 ## [0.13.0] - 2023-10-18
 ### Added